Main: the parser script was revised and redesigned for convenient interaction through the command console
