+++ KAIScheduleParser/.idea/KAIScheduleParser.iml
@@ -7,4 +7,7 @@
     <orderEntry type="jdk" jdkName="Python 3.11 (KAIScheduleParser)" jdkType="Python SDK" />
     <orderEntry type="sourceFolder" forTests="false" />
   </component>
+  <component name="PackageRequirementsSettings">
+    <option name="removeUnused" value="true" />
+  </component>
 </module>
